Always allocate notifiers
Previous versions of tachyonix would cache the notifiers and their
associated waker in the sender. While in theory this spares both on
allocation and waker cloning, it appears to prevent some optimizations
and actually results in worse performance across nearly all benchmarks
and benchmark parameters.

So this is a "worse" version which appears to perform better and makes
it possible to have the sender take `self` by shared reference rather
than by mutable reference.

Also, this commit removes the `UnsafeCell` around the waker, which was
there only to enable tracking through loom but which turned out to have
a non-negligible impact on performance.